final_project/requirements.txt
autopep8==1.7.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Flask==1.1.2
Flask-WTF==0.14.3
ibm-cloud-sdk-core==1.5.1
ibm-watson==4.4.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.1.1
pycodestyle==2.9.1
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==0.13.0
requests==2.28.1
six==1.16.0
toml==0.10.2
urllib3==1.26.12
websocket-client==0.48.0
Werkzeug==2.0.3
WTForms==3.0.1
